requirements.txt
# Core
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
tokenizers>=0.15.0
accelerate>=0.24.0

# Baseline
scikit-learn>=1.3.0

# LLM
anthropic>=0.39.0

# Evaluation
ragas>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.10

# Data & Viz
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
wordcloud>=1.9.0

# Config
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
